// agentark-web/src/shared/api/generated/control/models/environment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AgentArk Public Control API
 * OpenAPI spec version: 1.0.0
 */
import type { EnvironmentStatus } from "./environmentStatus";

export interface Environment {
  /** @pattern ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$ */
  id: string;
  /** @pattern ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$ */
  organizationId: string;
  /** @pattern ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$ */
  projectId: string;
  key: string;
  /**
   * @minLength 1
   * @maxLength 128
   */
  name: string;
  status: EnvironmentStatus;
  /** @minimum 0 */
  version: number;
  createdAt: string;
  updatedAt: string;
}